kubernetes/examples/kubernetes-engine-samples/cost-optimization/gke-vpa-recommendations/metrics-exporter/requirements.txt
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-monitoring
protobuf~=4.25.0
stackdriver_log_formatter